requirements.txt
torch
torchvision
numpy
Pillow
pydicom
opencv-python-headless
scikit-image
SimpleITK
matplotlib
pandas
